# docs/openapi.yml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Order
  description: Order and tickets management APIs
- name: Show
  description: Контроллер для работы с шоу
- name: Theatre
  description: Theatre management APIs
paths:
  /api/theatres:
    get:
      tags:
      - Theatre
      summary: Получить список театров в городе
      operationId: getAllTheatres
      parameters:
      - name: city
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCountDtoTheatreViewDto"
    put:
      tags:
      - Theatre
      summary: Обновить театр
      operationId: updateTheatre
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TheatrePayload"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TheatrePayload"
    post:
      tags:
      - Theatre
      summary: Создать театр
      operationId: createTheatre
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TheatrePayload"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TheatrePayload"
  /api/theatres/halls/{id}:
    put:
      tags:
      - Theatre
      summary: Обновить зал
      operationId: updateHall
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HallDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HallDto"
    delete:
      tags:
      - Theatre
      summary: Удалить зал
      operationId: deleteHall
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/shows/performance/{id}:
    put:
      tags:
      - Show
      summary: Обновить Performance
      operationId: updatePerformance
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformanceDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PerformanceDto"
  /api/theatres/{theatreId}/halls:
    post:
      tags:
      - Theatre
      summary: Создать зал для театра
      operationId: createHall
      parameters:
      - name: theatreId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HallDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HallDto"
  /api/shows/performance:
    post:
      tags:
      - Show
      summary: Создать Performance
      operationId: createPerformance
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformanceDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PerformanceDto"
  /api/orders/{orderId}/pay:
    post:
      tags:
      - Order
      summary: Оплата билетов по заказу
      operationId: payTickets
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDto"
  /api/orders/reserve:
    post:
      tags:
      - Order
      summary: Резервирование билетов
      operationId: reserveTickets
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPayload"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDto"
  /api/theatres/{id}:
    get:
      tags:
      - Theatre
      summary: Получить информацию о театре по ID
      operationId: getTheatreInfo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TheatreDto"
  /api/shows:
    get:
      tags:
      - Show
      summary: Получить список шоу в городе (с пагинацией)
      operationId: getAllShow
      parameters:
      - name: city
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCountDtoShowViewDto"
  /api/shows/{id}:
    get:
      tags:
      - Show
      summary: Получить информацию о шоу по id
      operationId: getShowInfo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShowDto"
  /api/shows/{id}/seats:
    get:
      tags:
      - Show
      summary: Получить список мест для шоу
      operationId: getAllSeats
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeatRawDto"
  /api/orders/{orderId}/tickets:
    get:
      tags:
      - Order
      summary: Получение билетов по заказу с постраничной прокруткой
      operationId: getTicketsPage
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageScrollDtoTicketDto"
components:
  schemas:
    HallViewDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
      required:
      - number
    TheatrePayload:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        address:
          type: string
        halls:
          type: array
          items:
            $ref: "#/components/schemas/HallViewDto"
      required:
      - address
      - city
      - halls
      - name
    HallDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        seatRows:
          type: array
          items:
            $ref: "#/components/schemas/SeatRawDto"
      required:
      - number
      - seatRows
    SeatRawDto:
      type: object
      properties:
        row:
          type: integer
          format: int32
        seats:
          type: array
          items:
            $ref: "#/components/schemas/SeatStatusDto"
      required:
      - row
      - seats
    SeatStatusDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - FREE
          - OCCUPIED
        number:
          type: integer
          format: int32
        price:
          type: integer
          format: int32
      required:
      - id
      - number
      - price
      - status
    PerformanceDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
      required:
      - title
    OrderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        reservedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - PAID
          - RESERVED
          - CANCELLED
        seat:
          type: array
          items:
            $ref: "#/components/schemas/SeatDto"
        price:
          type: integer
          format: int32
      required:
      - createdAt
      - id
      - price
      - seat
      - status
    SeatDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        raw:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
      required:
      - id
      - number
      - raw
    OrderPayload:
      type: object
      properties:
        seatIds:
          type: array
          items:
            type: integer
            format: int64
        showId:
          type: integer
          format: int64
      required:
      - seatIds
      - showId
    PageCountDtoTheatreViewDto:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/TheatreViewDto"
      required:
      - content
      - pageNumber
      - pageSize
    TheatreViewDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        address:
          type: string
      required:
      - address
      - city
      - id
      - name
    ShowViewDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tittle:
          type: string
        date:
          type: string
          format: date-time
      required:
      - date
      - id
      - tittle
    TheatreDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        address:
          type: string
        halls:
          type: array
          items:
            $ref: "#/components/schemas/HallViewDto"
        shows:
          type: array
          items:
            $ref: "#/components/schemas/ShowViewDto"
      required:
      - address
      - city
      - halls
      - id
      - name
      - shows
    PageCountDtoShowViewDto:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ShowViewDto"
      required:
      - content
      - pageNumber
      - pageSize
    ShowDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          format: int32
        hall:
          $ref: "#/components/schemas/HallViewDto"
        theatre:
          $ref: "#/components/schemas/TheatreViewDto"
      required:
      - date
      - description
      - hall
      - id
      - theatre
      - title
    PageScrollDtoTicketDto:
      type: object
      properties:
        hasNextPage:
          type: boolean
        content:
          type: array
          items:
            $ref: "#/components/schemas/TicketDto"
      required:
      - content
      - hasNextPage
    TicketDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        price:
          type: integer
          format: int32
        raw:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
      required:
      - id
      - number
      - price
      - raw